<feat>(project): 做出以下改动
1.修改led_frame中.h中呼吸周期位数为16
2.修改ws2812_led_bar.h中呼吸次数位数为16/rgb步进量为float类型
  新增单次呼吸周期变量 用于统一数值
3.修改drv_i2c.c中control_i2c()函数I2C_GET_RECV_BUFF方法实现
4.修改led_bar.c中呼吸驱动回调函数参数顺序及数据分析函数中数据解析逻辑
5.调整task_sch.c中任务调度函数内 自主定义任务的优先级为最高
6.自测 新增静态颜色功能正常 drv_i2c设备驱动正常 以及呼吸模式正常
7.呼吸模式：有时存在颜色抖动情况 怀疑是电压不稳定导致的

+++ Template/platform/Inc/led_frame.h
@@ -43,7 +43,7 @@ struct led_bar{
     Rtv_Status (*off)(struct led_bar *bar );
     Rtv_Status (*blink)(struct led_bar *bar, uint8_t mode, uint8_t blink_led_num);
     Rtv_Status (*water)(struct led_bar *bar, uint8_t mode, uint8_t single_led_num, uint8_t move_period);
-    Rtv_Status (*breath)(struct led_bar *bar, uint8_t breath_period);
+    Rtv_Status (*breath)(struct led_bar *bar, uint16_t breath_period);
     void *private;
 };
 typedef struct led_bar* led_bar_t;
